have the "create_sam_args" in evaulation to a seperate util file so train can use it as well
