+++ train.py
@@ -6,6 +6,7 @@ from tqdm import tqdm
 import os
 import json
 import numpy as np
+from utils.sam_utils import get_sam_args
 from models.model_single import ModelEmb
 from dataset.glas import get_glas_dataset
 from dataset.MoNuBrain import get_monu_dataset
@@ -321,6 +322,10 @@ if __name__ == '__main__':
     parser.add_argument('--debug', action='store_true', help='debug mode', required=False)
     parser.add_argument('--run_name', help='run name, its mendaotry as we need easy way to tell between runs',
                          required=True)
+    parser.add_argument('--sam_model_type', type=str,
+                    default='vit_b', help='SAM model type', required=False)
+    parser.add_argument('--sam_checkpoint_dir_path', type=str, default='cp',
+                        help='SAM checkpoint directory path', required=False)
     args = vars(parser.parse_args())
     run_output_path = os.path.join('results', args['run_name'])
     args['run_output_path'] = run_output_path
@@ -333,36 +338,10 @@ if __name__ == '__main__':
                                      'net_best.pth')
     args['vis_folder'] = os.path.join(run_output_path, 'vis')
     os.makedirs(args['vis_folder'], exist_ok=True)
-    # sam_args = {
-    #     'sam_checkpoint': "cp/sam_vit_h.pth",
-    #     'model_type': "vit_h",
-    #     'generator_args': {
-    #         'points_per_side': 8,
-    #         'pred_iou_thresh': 0.95,
-    #         'stability_score_thresh': 0.7,
-    #         'crop_n_layers': 0,
-    #         'crop_n_points_downscale_factor': 2,
-    #         'min_mask_region_area': 0,
-    #         'point_grids': None,
-    #         'box_nms_thresh': 0.7,
-    #     },
-    #     'gpu_id': 0,
-    # }
-    sam_args = {
-        'sam_checkpoint': "cp/sam_vit_b.pth",
-        'model_type': "vit_b",
-        'generator_args': {
-            'points_per_side': 8,
-            'pred_iou_thresh': 0.95,
-            'stability_score_thresh': 0.7,
-            'crop_n_layers': 0,
-            'crop_n_points_downscale_factor': 2,
-            'min_mask_region_area': 0,
-            'point_grids': None,
-            'box_nms_thresh': 0.7,
-        },
-        'gpu_id': 0,
-    }
+
+    sam_args = get_sam_args(sam_model_type=args["sam_model_type"],
+                            sam_checkpoint_dir_path=args["sam_checkpoint_dir_path"])
+
     # save to json file all the info of the current run 
     run_info = {"run_name": args['run_name']}
     run_info['args'] = args
